refactor: add js doc volunteer controller

+++ src/modules/volunteers/VolunteerController.ts
@@ -20,14 +20,21 @@ class VolunteerController extends Controller {
   /**
    * @description Create new volunteer
    *
-   * @method POST
    * @route /volunteers
+   * @method POST
    * @access private
+   *
+   * @param {Request} req - The request object volunteer details payload.
+   * @param {Response} res - The response object used to send the response.
+   * @returns {Response} A response indicating the created volunteer details.
+   * @throws {AppError} If the volunteer email already exists.
+   * @throws {AppError} If the volunteer phone number already exists.
+   * @throws {AppError} If error occurred during volunteer creation
    */
   static createVolunteer: RequestHandler = async (
     req: Request,
     res: Response
-  ) => {
+  ): Promise<Response<any, Record<string, any>>> => {
     const payload: IVolunteerPayload = req.body;
 
     if (payload.email) {
@@ -81,14 +88,19 @@ class VolunteerController extends Controller {
   /**
    * @description Get volunteer details by ID
    *
-   * @method GET
    * @route /volunteers/:volunteer_id
+   * @method GET
    * @access private
+   *
+   * @param {Request} req - The request object volunteer id.
+   * @param {Response} res - The response object used to send the response.
+   * @returns {Response} A response indicating the created volunteer details.
+   * @throws {AppError} If the volunteer details not found.
    */
   static getVolunteerById: RequestHandler = async (
     req: Request,
     res: Response
-  ) => {
+  ): Promise<Response<any, Record<string, any>>> => {
     const volunteerId: string = req.params.volunteer_id;
 
     const existVolunteer = await VolunteerService.getVolunteerDetails(
@@ -115,16 +127,21 @@ class VolunteerController extends Controller {
   };
 
   /**
-   *  @description Get all volunteers
+   * @description Get all volunteers
    *
+   * @route /volunteers
    * @method GET
-   * @route /volunteers/:volunteer_id
    * @access private
+   *
+   * @param {Request} req - The request object.
+   * @param {Response} res - The response object used to send the response.
+   * @returns {Response} A response indicating the retrieved volunteers.
+   * @throws {AppError} If no volunteer details found.
    */
   static getAllVolunteers: RequestHandler = async (
     req: Request,
     res: Response
-  ) => {
+  ): Promise<Response<any, Record<string, any>>> => {
     const { page, limit } = req.pagination;
     const { name } = req.query;
     let response: IVolunteerReponse[] = [];
@@ -184,14 +201,22 @@ class VolunteerController extends Controller {
   /**
    * @description Update volunteer details
    *
-   * @method PUT
    * @route /volunteers/:volunteer_id
+   * @method PUT
    * @access private
+   *
+   * @param {Request} req - The request object containing volunteer details.
+   * @param {Response} res - The response object used to send the response.
+   * @returns {Response} - A response indicating the updated volunteer.
+   * @throws {AppError} - If selected volunteer details not found.
+   * @throws {AppError} - If selected volunteer details duplicates email.
+   * @throws {AppError} - If selected volunteer details duplicates phone number.
+   * @throws {AppError} - If failed to update volunteer details.
    */
   static updateVolunteerDetails: RequestHandler = async (
     req: Request,
     res: Response
-  ) => {
+  ): Promise<Response<any, Record<string, any>>> => {
     const volunteerId: string = req.params.volunteer_id;
     const payload: IVolunteerPayload = req.body;
 
@@ -258,16 +283,23 @@ class VolunteerController extends Controller {
   };
 
   /**
-   * @description Delete volunteer details
+   * @description Delete volunteer details by ID
    *
-   * @method DELETE
    * @route /volunteers/:volunteer_id
+   * @method DELETE
    * @access private
+   *
+   * @param {Request} req - The request object volunteer id.
+   * @param {Response} res - The response object used to send the response.
+   * @returns {Response} A response indicating the volunteer details deleted
+   * @throws {AppError} If the volunteer details not found.
+   * @throws {AppError} If volunteer is active.
+   * @throws {AppError} If failed to delete volunteer details.
    */
   static deleteVolunteer: RequestHandler = async (
     req: Request,
     res: Response
-  ) => {
+  ): Promise<Response<any, Record<string, any>>> => {
     const volunteerId: string = req.params.volunteer_id;
 
     const existVolunteer = await VolunteerService.getVolunteerDetails(
@@ -281,12 +313,12 @@ class VolunteerController extends Controller {
     }
 
     // do not delete active volunteer
-    // if (existVolunteer.status) {
-    //   throw new AppError({
-    //     httpCode: HttpCode.BAD_REQUEST,
-    //     description: "Active volunteer cannot be deleted",
-    //   });
-    // }
+    if (existVolunteer.status) {
+      throw new AppError({
+        httpCode: HttpCode.BAD_REQUEST,
+        description: "Active volunteer cannot be deleted",
+      });
+    }
 
     const deletedVolunteer = await VolunteerService.deleteVolunteerDetails(
       volunteerId
